requirements.txt
Click==7.0
cycler==0.10.0
datefinder==0.7.0
decorator==4.4.1
Flask==1.1.1
gunicorn==20.0.4
imageio==2.6.1
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
jsonpickle==1.2
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
networkx==2.4
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
pyparsing==2.4.5
pytesseract==0.3.0
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
regex==2019.11.1
scikit-image==0.16.2
scikit-learn==0.22
scipy==1.3.3
six==1.13.0
sklearn==0.0
Werkzeug==0.16.0
